#192: wire real FileClient sender for GUI/Android send; use peer IP not reverse-DNS host
The base AppContainer.batchSenderFactory was a throw-stub (PeerUnreachableException);
only the CLI container overrode it with a real sender, so Desktop-GUI / Android / iOS
sends always failed with 'peer no longer reachable'. Wire the real FileClient-backed
sender into the base container (dedupes the CLI's identical buildCliBatchSender).

Also: the /hello rendezvous upsert stored call.request.origin.remoteHost, which Ktor
reverse-resolves to a hostname (e.g. 'Mac') the sender platform cannot resolve back to
an IP — switch to remoteAddress (numeric IP), matching the existing HelloRouteTest intent.

Verified on-device: Android picked a 2.15 MB image via SAF and it landed byte-complete
on a Desktop peer.
