+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/di/AppContainer.kt
@@ -23,6 +23,7 @@ import com.tubetoast.tether.preferences.PeerPreferencesStore
 import com.tubetoast.tether.presentation.RootComponentFactory
 import com.tubetoast.tether.presentation.banners.PeerConflictRelay
 import com.tubetoast.tether.protocol.DeviceType
+import com.tubetoast.tether.protocol.SendResult
 import com.tubetoast.tether.security.DefaultTrustedDeviceStore
 import com.tubetoast.tether.security.TrustedDeviceStore
 import com.tubetoast.tether.transfer.AutoSendDispatcher
@@ -36,6 +37,7 @@ import com.tubetoast.tether.transfer.PeerTransferEngineRegistry
 import com.tubetoast.tether.transfer.PeerUnreachableException
 import com.tubetoast.tether.transfer.PendingFilesRepository
 import com.tubetoast.tether.transfer.ReconnectionTimeout
+import com.tubetoast.tether.transfer.toPeerIdentity
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -74,9 +76,30 @@ abstract class AppContainer {
 
     // Factory: BatchSender holds per-transfer state — one instance per concurrent peer transfer.
     open val batchSenderFactory: (PeerIdentity) -> BatchSender by lazy {
-        { _ ->
+        { peer ->
             BatchSender(
-                sendOne = { _, _ -> throw PeerUnreachableException() },
+                sendOne = { source, onProgress ->
+                    val device = discoveredDevicesStore.devices.value
+                        .firstOrNull { it.toPeerIdentity() == peer }
+                        ?: throw PeerUnreachableException()
+                    try {
+                        // FileClient.send folds all failures into SendResult.Failure — re-raise as typed.
+                        when (
+                            val result = fileClient.send(
+                                device = device,
+                                channel = source.openReadChannel(),
+                                fileName = source.name,
+                                totalBytes = source.sizeBytes,
+                                onProgress = onProgress,
+                            )
+                        ) {
+                            is SendResult.Success -> Unit
+                            is SendResult.Failure -> throw PeerUnreachableException(RuntimeException(result.reason))
+                        }
+                    } finally {
+                        source.close()
+                    }
+                },
                 connectionMonitor = connectionMonitor,
             )
         }
